+++ servers/front/main.go
@@ -10,8 +10,8 @@ import (
 	"github.com/ministryofjustice/opg-reports/servers/front/aws_costs"
 	"github.com/ministryofjustice/opg-reports/servers/front/config"
 	"github.com/ministryofjustice/opg-reports/servers/front/dl"
-	"github.com/ministryofjustice/opg-reports/servers/front/front_templates"
 	"github.com/ministryofjustice/opg-reports/servers/front/github_standards"
+	"github.com/ministryofjustice/opg-reports/servers/front/template_helpers"
 	"github.com/ministryofjustice/opg-reports/shared/env"
 	"github.com/ministryofjustice/opg-reports/shared/logger"
 )
@@ -42,7 +42,7 @@ func main() {
 	})
 
 	// -- get templates
-	templates = front_templates.GetTemplates("./templates")
+	templates = template_helpers.GetTemplates("./templates")
 	for _, f := range templates {
 		slog.Debug("template file", slog.String("path", f))
 	}
