+++ servers/front/template_helpers/template_helpers.go
@@ -1,4 +1,4 @@
-package front_templates
+package template_helpers
 
 import (
 	"os"
